:hammer: Abstract ResourcesClient implementation for multiplatform reuse
