+++ app/src/desktopMain/kotlin/com/crosspaste/net/DesktopResourcesClient.kt
@@ -1,33 +1,25 @@
 package com.crosspaste.net
 
-import com.crosspaste.app.AppFileType
 import com.crosspaste.config.DesktopConfigManager
 import com.crosspaste.path.UserDataPathProvider
 import com.crosspaste.ui.extension.ProxyType
 import com.crosspaste.utils.getFileUtils
 import io.github.oshai.kotlinlogging.KLogger
 import io.github.oshai.kotlinlogging.KotlinLogging
-import io.ktor.client.HttpClient
-import io.ktor.client.engine.ProxyBuilder
-import io.ktor.client.engine.cio.CIO
-import io.ktor.client.engine.http
-import io.ktor.client.plugins.logging.Logger
-import io.ktor.client.plugins.logging.Logging
-import io.ktor.client.plugins.onDownload
-import io.ktor.client.plugins.timeout
-import io.ktor.client.request.prepareRequest
+import io.ktor.client.*
+import io.ktor.client.engine.*
+import io.ktor.client.engine.cio.*
+import io.ktor.client.plugins.*
+import io.ktor.client.plugins.logging.*
 import io.ktor.client.request.request
-import io.ktor.client.statement.HttpResponse
-import io.ktor.client.statement.bodyAsChannel
+import io.ktor.client.statement.*
 import io.ktor.http.headers
-import io.ktor.http.isSuccess
 import io.ktor.util.collections.*
-import okio.Path
 
 class DesktopResourcesClient(
     val configManager: DesktopConfigManager,
-    val userDataPathProvider: UserDataPathProvider,
-) : ResourcesClient {
+    userDataPathProvider: UserDataPathProvider,
+) : AbstractResourcesClient(userDataPathProvider) {
 
     companion object {
 
@@ -59,7 +51,7 @@ class DesktopResourcesClient(
             }
     }
 
-    private val logger = KotlinLogging.logger {}
+    override val logger = KotlinLogging.logger {}
 
     private val noProxyClient = createClient(logger)
 
@@ -76,7 +68,7 @@ class DesktopResourcesClient(
         }
     }
 
-    private fun getHttpClient(): HttpClient {
+    override fun getHttpClient(): HttpClient {
         val proxy = getProxy()
         return proxy?.let {
             proxyClientMap.getOrPut(proxy) {
@@ -85,61 +77,7 @@ class DesktopResourcesClient(
         } ?: noProxyClient
     }
 
-    private fun getTempFilePath(): Path =
-        userDataPathProvider.resolve(
-            fileName = fileUtils.createRandomFileName(),
-            appFileType = AppFileType.TEMP,
-        )
-
-    override suspend fun download(
-        url: String,
-        path: Path,
-        listener: DownloadProgressListener,
-    ) {
-        var shouldShowProgress = false
-        getHttpClient()
-            .prepareRequest(url) {
-                onDownload { bytesSent, contentLength ->
-                    if (shouldShowProgress) {
-                        listener.onProgress(bytesSent, contentLength)
-                    }
-                }
-            }.execute { response ->
-                if (response.status.isSuccess()) {
-                    shouldShowProgress = true
-                    val tempFilePath = getTempFilePath()
-                    runCatching {
-                        val channel = response.bodyAsChannel()
-                        fileUtils.writeFile(tempFilePath, channel)
-                        fileUtils.moveFile(tempFilePath, path)
-                    }.onSuccess {
-                        listener.onSuccess()
-                    }.onFailure {
-                        if (fileUtils.existFile(tempFilePath)) {
-                            fileUtils.deleteFile(tempFilePath)
-                        }
-                        if (fileUtils.existFile(path)) {
-                            fileUtils.deleteFile(path)
-                        }
-                        listener.onFailure(response.status, it)
-                    }
-                } else {
-                    listener.onFailure(response.status, null)
-                }
-            }
-    }
-
-    override suspend fun request(url: String): Result<ClientResponse> {
-        val response = clientRequest(url, getHttpClient())
-        return if (response.status.isSuccess()) {
-            Result.success(ClientResponse(response))
-        } else {
-            logger.warn { "Failed to fetch data from $url, status code: ${response.status.value}" }
-            Result.failure(kotlin.Exception("HTTP error: ${response.status.value}"))
-        }
-    }
-
-    private suspend fun clientRequest(
+    override suspend fun clientRequest(
         url: String,
         client: HttpClient,
     ): HttpResponse =
